requirements.txt
alembic==1.11.1
Babel==2.10.3
blinker==1.6.2
Bootstrap-Flask==2.1.0
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cryptography==38.0.1
dnspython==2.2.1
Flask==2.3.2
Flask-Babel==2.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-Table==0.5.0
Flask-WTF==1.0.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
numpy==1.25.0
pandas==2.0.3
phonenumbers==8.12.56
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.17
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.4
uuid==1.30
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.9.0